+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.1.0 2021-01-22
+
+- Added support for labyrinths with circular paths
+- Added support for "empty" labyrinths (only outer walls and cheese)
+
 ## 1.0.4 2021-01-22
 
 - First fully functional release